# src/Spotless.Web/lib/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Spotless API",
    "description": "Laundry and dry cleaning service API",
    "version": "v1"
  },
  "paths": {
...
    "/api/customers/cart/items": {
      "post": {
        "tags": [
          "Carts"
        ],
        "summary": "Adds a service to customer's cart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Cart.AddToCartDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Cart.AddToCartDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Cart.AddToCartDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/customers/cart/items/{serviceId}": {
      "delete": {
        "tags": [
          "Carts"
        ],
        "summary": "Removes a service from customer's cart",
        "parameters": [
          {
            "name": "serviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/customers/cart/checkout": {
      "post": {
        "tags": [
          "Carts"
        ],
        "summary": "Checks out cart and creates an order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Cart.CartCheckoutRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Cart.CartCheckoutRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Cart.CartCheckoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/customers/cart/buy-now": {
      "post": {
        "tags": [
          "Carts"
        ],
        "summary": "Creates immediate order without cart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Cart.BuyNowRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Cart.BuyNowRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Cart.BuyNowRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Lists all service categories with pagination",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Responses.PagedResponse`1[[Spotless.Application.Dtos.Category.CategoryDto, Spotless.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Responses.PagedResponse`1[[Spotless.Application.Dtos.Category.CategoryDto, Spotless.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Responses.PagedResponse`1[[Spotless.Application.Dtos.Category.CategoryDto, Spotless.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Creates a new category",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Category.CreateCategoryDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Category.CreateCategoryDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Category.CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Category.CategoryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Category.CategoryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Category.CategoryDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/categories/{id}": {
      "put": {
        "tags": [
          "Categories"
        ],
        "summary": "Updates an existing category",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Category.UpdateCategoryDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Category.UpdateCategoryDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Spotless.Application.Dtos.Category.UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Category.CategoryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Category.CategoryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spotless.Application.Dtos.Category.CategoryDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "Deletes a category",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "Spotless.Application.Dtos.Cart.AddToCartDto": {
        "type": "object",
        "properties": {
          "serviceId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Spotless.Application.Dtos.Cart.BuyNowRequest": {
        "type": "object",
        "properties": {
          "serviceId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "timeSlotId": {
            "type": "string",
            "format": "uuid"
          },
          "scheduledDate": {
            "type": "string",
            "format": "date-time"
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/Spotless.Domain.Enums.PaymentMethod"
          },
          "pickupLatitude": {
            "type": "number",
            "format": "double"
          },
          "pickupLongitude": {
            "type": "number",
            "format": "double"
          },
          "pickupAddress": {
            "type": "string",
            "nullable": true
          },
          "deliveryLatitude": {
            "type": "number",
            "format": "double"
          },
          "deliveryLongitude": {
            "type": "number",
            "format": "double"
          },
          "deliveryAddress": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Spotless.Application.Dtos.Cart.CartCheckoutRequest": {
        "type": "object",
        "properties": {
          "timeSlotId": {
            "type": "string",
            "format": "uuid"
          },
          "scheduledDate": {
            "type": "string",
            "format": "date-time"
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/Spotless.Domain.Enums.PaymentMethod"
          },
          "pickupLatitude": {
            "type": "number",
            "format": "double"
          },
          "pickupLongitude": {
            "type": "number",
            "format": "double"
          },
          "pickupAddress": {
            "type": "string",
            "nullable": true
          },
          "deliveryLatitude": {
            "type": "number",
            "format": "double"
          },
          "deliveryLongitude": {
            "type": "number",
            "format": "double"
          },
          "deliveryAddress": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Spotless.Application.Dtos.Category.CategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "serviceCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Spotless.Application.Dtos.Category.CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "Spotless.Application.Dtos.Category.UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "Spotless.Domain.Enums.PaymentMethod": {
        "enum": [
          "CreditCard",
          "DebitCard",
          "Fawry",
          "PayPal",
          "Wallet",
          "CashOnDelivery",
          "Paymob",
          "Instapay"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}